% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heavyRain-package.R
\docType{data}
\name{baleCHIRPS.v2}
\alias{baleCHIRPS.v2}
\title{Bale Mountains CHIRPS.v2}
\format{
\code{raster::RasterStack}
}
\description{
Bale Mountains CHIRPS.v2.
}
\details{
This dataset contains CHIRPS.v2 observations for the Bale Mountains
National Park, southern Ethiopia (Jan 1981 to Dec 2016).
}
