% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadChirps.R
\name{downloadChirps}
\alias{downloadChirps}
\title{Download CHIRPS data}
\arguments{
\item{region}{'character. Region of interest; one of 'global', 'whem'
(western hemisphere), 'africa' or 'camer-carib' (Central America and the
Caribbean). See the official CHIRPS README (available online at
\url{ftp://chg-ftpout.geog.ucsb.edu/pub/org/chg/products/CHIRPS-2.0/README-CHIRPS.txt})
for further information.}

\item{format}{'character'. Desired file format; one of 'tifs', 'netcdf',
'bils' or 'pngs'.}

\item{tres}{'character'. Desired temporal resolution; one of 'daily',
'pentad', 'dekad' or 'monthly'.}

\item{sres}{'numeric'. Desired spatial resolution; one of
\code{c(0.05, 0.25)}.}

\item{begin}{'date'. If not supplied, data download starts with the first
record available.}

\item{end}{'date'. If not supplied, data download stops at the last record
available.}

\item{dsn}{'character'. Destination folder for file download, defaults to the
current working directory.}

\item{overwrite}{Logical. If \code{TRUE}, already downloaded files in 'dsn'
will be overwritten.}

\item{cores}{'integer'. The number of cores used for parallel download via
\strong{doParallel}.}

\item{...}{Further arguments passed on to \code{\link{download.file}}.}
}
\value{
A vector of filepaths.
}
\description{
Download Climate Hazards Group InfraRed Precipitation with Station (CHIRPS)
data for a given time span from Climate Hazards Group's (CHG) FTP server
(\url{ftp://chg-ftpout.geog.ucsb.edu/pub/org/chg/products/CHIRPS-2.0/}).
}
\examples{
\dontrun{
## Download GIMMS NDVI3g binary data from 2000-2005 (this might take some time...)
gimms_files <- downloadGimms(x = 2000, y = 2005,
                             dsn = paste0(getwd(), "/data"))
gimms_files[1:10]
}
}
\author{
Florian Detsch
}
\seealso{
\code{\link{download.file}}
}

