% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractChirps.R
\name{extractChirps}
\alias{extractChirps}
\title{Extract compressed CHIRPS files}
\usage{
extractChirps(x, exdir, cores = 1L, ...)
}
\arguments{
\item{x}{'character'. Input filename(s).}

\item{exdir}{'character'. The destination folder for file extraction.}

\item{cores}{'integer'. The number of cores used for parallel processing via
\strong{doParallel}.}

\item{...}{Further arguments passed on to \code{\link{gunzip}}, e.g. 'skip',
'overwrite' and 'remove'.}
}
\value{
A 'character' vector of filenames.
}
\description{
This wrapper function around \code{\link{gunzip}} extracts the contents of
a single or multiple CHIRPS *.gz files to a specified destination folder.
}
\author{
Florian Detsch
}

