% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heavyRain-package.R
\docType{package}
\name{heavyRain-package}
\alias{heavyRain-package}
\alias{heavyRainpackage}
\title{Download and Pre-Process CHIRPS and TRMM Rainfall Data Sets in R}
\description{
Download and Pre-Process CHIRPS and TRMM Rainfall Data Sets in R
}
\details{
This is a set of functions to download and pre-process CHIRPS and TRMM
rainfall data sets in R.
}
\author{
Florian Detsch
}
\keyword{package}
