% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyIndices.R
\name{monthlyIndices}
\alias{monthlyIndices}
\title{Create timestamps or monthly indices from CHIRPS files}
\usage{
monthlyIndices(x, pos1 = 13L, pos2 = 22L, timestamp = FALSE, ...)
}
\arguments{
\item{x}{'character'. Input filename(s).}

\item{pos1, pos2}{'integer'. The first and last element of the date string in
'x', defaults to the CHIRPS naming convention; see \code{\link{substr}}.}

\item{timestamp}{'logical'. If TRUE, an actual timestamp (formatted according
to \code{...}) is returned rather than a vector of indices.}

\item{...}{Further arguments passed on to \code{\link{strftime}}.}
}
\value{
An 'integer' vector with unique monthly indices or, if
\code{timestamp = TRUE}, a character vector with formatted timestamps.
}
\description{
Extract time information or, based upon this, monthly indices from regularly
named CHIRPS files (compressed or decompressed).
}
\author{
Florian Detsch
}

