% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedMannKendall.R
\name{stratifiedMannKendall}
\alias{stratifiedMannKendall}
\title{Compute stratified Mann-Kendall trend tests}
\usage{
stratifiedMannKendall(x, type = c("raw", "monthly", "seasonal"), p = 0.001,
  prewhitening = TRUE, cores = 1L)
}
\arguments{
\item{x}{A 'RasterStack' consisting of consecutive time series observations.}

\item{type}{'character'; one of 'raw' (take the input series as is),
'monthly' (Mann-Kendall test across unique months) and 'seasonal'
(Mann-Kendall test across unique seasons).}

\item{p}{Significance level to be tested.}

\item{prewhitening}{'logical'. If \code{TRUE}, pre-whitening is applied prior
to the actual Mann-Kendall trend test. See also \code{\link{significantTau}}
and description of package \strong{zyp}
(\url{https://cran.r-project.org/web/packages/zyp/zyp.pdf}).}

\item{cores}{If desired number of cores to use for parallel processing.}
}
\value{
A 'list' with single entries corresponding to the significance levels in
\code{p}.
}
\description{
In contrast to most existing implementation, this function allows you to
apply pixel-based Mann-Kendall trend tests not only to a raw 'RasterStack' of
time series observations, but also to stratified subsets of the very same,
i.e., across monthly (Jan-Dec) and seasonal ('DJF', 'MAM', 'JJA', 'SON')
chunks of layers.
}
\author{
Florian Detsch
}
\seealso{
\code{\link{significantTau}}.
}

