% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRain.R
\name{getRain}
\alias{getRain}
\title{Download Rainfall Data Set}
\usage{
getRain(name = c("chirps", "trmm"), ...)
}
\arguments{
\item{name}{\code{character}, currently one of \code{"chirps"} (default) or
\code{"trmm"}.}

\item{...}{Additional arguments passed to underlying functions listed under
See Also.}
}
\value{
A \code{character} vector of file paths.
}
\description{
Download a particular rainfall data set. Currently available products
originate from Climate Hazards Group InfraRed Precipitation with Station Data
(CHIRPS, \url{http://chg.geog.ucsb.edu/data/chirps/}) and the Tropical
Rainfall Measuring Mission (TRMM, \url{https://pmm.nasa.gov/trmm}).
}
\seealso{
\code{\link{getCHIRPS}}, \code{\link{getTRMM}}.
}
