% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedMannKendall.R
\name{stratifiedMannKendall}
\alias{stratifiedMannKendall}
\title{Compute stratified Mann-Kendall trend tests}
\usage{
stratifiedMannKendall(x, type = c("raw", "monthly", "seasonal"), p = 0.001,
  prewhitening = TRUE, cores = 1L)
}
\arguments{
\item{x}{A 'RasterStack' consisting of consecutive time series observations.}

\item{type}{'character'; one of 'raw' (take the input series as is),
'monthly' (Mann-Kendall test across unique months) and 'seasonal'
(Mann-Kendall test across unique seasons).}

\item{p}{Significance level to be tested.}

\item{prewhitening}{'logical'. If \code{TRUE}, pre-whitening is applied prior
to the actual Mann-Kendall trend test.}

\item{cores}{Number of cores for parallel processing.}
}
\value{
A 'Raster*' object.
}
\description{
In contrast to most existing implementations, this function allows you to
apply pixel-based Mann-Kendall trend tests not only to a raw 'RasterStack' of
time series observations, but also to stratified subsets of the very same,
i.e., across monthly (Jan-Dec) and seasonal ('DJF', 'MAM', 'JJA', 'SON')
chunks of layers.
}
\examples{
\dontrun{
## overall trend
overall_trend = stratifiedMannKendall(baleCHIRPS.v2, p = 1)
spplot(overall_trend)

## monthly trend
monthly_trend = stratifiedMannKendall(baleCHIRPS.v2, type = "monthly", p = 0.05)
names(monthly_trend) = month.abb
spplot(monthly_trend)

## seasonal trend
seasons_trend = stratifiedMannKendall(baleCHIRPS.v2, type = "seasonal", p = 0.05)
names(seasons_trend) = c("DJF", "MAM", "JJA", "SON")
spplot(seasons_trend)
}

}
\seealso{
\code{\link{significantTau}}, \code{\link{confint}}.
}
\author{
Florian Detsch
}
